+++ src/HeaderCell.tsx
@@ -22,9 +22,13 @@ export interface HeaderCellProps<R, SR> extends SharedHeaderRowProps<R, SR> {
 
 export default function HeaderCell<R, SR>({
   column,
+  lastFrozenColumnIndex,
+  onResize,
   allRowsSelected,
   onAllRowsSelectionChange,
-  ...props
+  sortColumn,
+  sortDirection,
+  onSort
 }: HeaderCellProps<R, SR>) {
   function getCell() {
     if (!column.headerRenderer) return column.name;
@@ -38,9 +42,9 @@ export default function HeaderCell<R, SR>({
     cell = (
       <SortableHeaderCell
         column={column}
-        onSort={props.onSort}
-        sortColumn={props.sortColumn}
-        sortDirection={props.sortDirection}
+        onSort={onSort}
+        sortColumn={sortColumn}
+        sortDirection={sortDirection}
       >
         {cell}
       </SortableHeaderCell>
@@ -49,7 +53,7 @@ export default function HeaderCell<R, SR>({
 
   const className = classNames('rdg-cell', column.headerCellClass, {
     'rdg-cell-frozen': column.frozen,
-    'rdg-cell-frozen-last': column.idx === props.lastFrozenColumnIndex
+    'rdg-cell-frozen-last': column.idx === lastFrozenColumnIndex
   });
   const style: React.CSSProperties = {
     width: column.width,
@@ -69,7 +73,7 @@ export default function HeaderCell<R, SR>({
     cell = (
       <ResizableHeaderCell
         column={column}
-        onResize={props.onResize}
+        onResize={onResize}
       >
         {cell as React.ReactElement<React.ComponentProps<'div'>>}
       </ResizableHeaderCell>