+++ src/RowRenderer.tsx
@@ -18,7 +18,6 @@ interface IRowRendererProps<R, SR> extends SharedDataGridProps<R, SR> {
   lastFrozenColumnIndex: number;
   rowIdx: number;
   row: R;
-  enableCellRangeSelection?: boolean;
   eventBus: EventBus;
   isRowSelected: boolean;
 }
@@ -31,18 +30,19 @@ function RowRenderer<R, SR>({
   row,
   rowGroupRenderer,
   rowRenderer,
-  ...props
+  isRowSelected,
+  onRowClick,
+  onRowExpandToggle
 }: IRowRendererProps<R, SR>) {
   const { __metaData } = row as RowData;
   const rendererProps: RowRendererProps<R, SR> = {
     rowIdx,
     row,
     viewportColumns,
-    isRowSelected: props.isRowSelected,
+    isRowSelected,
     lastFrozenColumnIndex,
     eventBus,
-    onRowClick: props.onRowClick,
-    enableCellRangeSelection: props.enableCellRangeSelection
+    onRowClick
   };
 
   if (__metaData) {
@@ -57,7 +57,7 @@ function RowRenderer<R, SR>({
           name={(row as RowData).name!}
           eventBus={eventBus}
           renderer={rowGroupRenderer}
-          onRowExpandToggle={props.onRowExpandToggle}
+          onRowExpandToggle={onRowExpandToggle}
         />
       );
     }